Move `createDropdownItem` to `github-helpers/`
